+++ oneflow/python/nn/modules/math_ops.py
@@ -958,36 +958,3 @@ def pow_op(tensor, exponent):
         
     """
     return Pow()(tensor, exponent)
-
-
-class Log1p(Module):
-    def __init__(self) -> None:
-        super().__init__()
-        self._op = flow.builtin_op("log1p").Input("x").Output("y").Build()
-
-    def forward(self, x):
-        return self._op(x)[0]
-
-
-@oneflow_export("log1p")
-@register_tensor_op("log1p")
-@experimental_api
-def log1p_op(input):
-    r"""Returns a new tensor with the natural logarithm of (1 + input).
-    
-    .. math::
-        \text{out}_{i}=\log_e(1+\text{input}_{i})
-
-    For example:
-
-    .. code-block:: python
-
-        import oneflow.experimental as flow
-        import numpy as np
-        
-        x = flow.Tensor(np.array([[1.3, 1.5, 2.7]))
-        out = flow.log1p(x).numpy()
-        print(out) # [0.8329091  0.91629076 1.3083328]
-        
-    """
-    return Log1p()(input)
